refactor: setErrorMessag의 이름을 함수명과 동일하게 replaceAt으로 수정
